# Conflicts:
#	common/src/commonMain/kotlin/com/artemchep/keyguard/di/GlobalModuleCommon.kt
#	common/src/commonMain/kotlin/com/artemchep/keyguard/feature/home/settings/SettingPaneContent.kt
#	common/src/desktopMain/kotlin/com/artemchep/keyguard/core/session/PlatformApplicationModule.kt
#	desktopApp/build.gradle.kts
#	desktopApp/src/jvmMain/kotlin/com/artemchep/keyguard/Main.kt

+++ util/kdbx/src/commonMain/kotlin/app/keemobile/kotpass/database/modifiers/Content.kt
@@ -50,7 +50,9 @@ inline fun KeePassDatabase.withRecycleBin(
  * @return modified [KeePassDatabase].
  */
 fun KeePassDatabase.cleanupHistory(reference: Instant = Clock.System.now()): KeePassDatabase {
-    val maintenancePeriod = content.meta.maintenanceHistoryDays.toInt().days
+    // Widen to Long: values above Int.MAX_VALUE (e.g. UInt.MAX_VALUE used as
+    // "keep forever") would otherwise wrap negative and drop all history.
+    val maintenancePeriod = content.meta.maintenanceHistoryDays.toLong().days
 
     return modifyContent {
         copy(
